move constants folder in to @ folder
