+++ imports/ui/LayoutContainer.jsx
@@ -38,7 +38,7 @@ export const StateContext = React.createContext(null);
 import UserPopup from './components/UserPopup';
 import FormField from './components/FormField';
 import Hosts from '../api/@hosts/host';
-import { chakraTheme } from './constants/theme';
+import { chakraTheme } from './@/constants/theme';
 
 const publicSettings = Meteor.settings.public;
