+++ imports/ui/pages/admin/Emails.jsx
@@ -14,13 +14,13 @@ import { Controller, useForm } from 'react-hook-form';
 
 import Template from '../../components/Template';
 import ListMenu from '../../components/ListMenu';
-import { editorFormats, editorModules } from '../../constants/quillConfig';
+import { editorFormats, editorModules } from '../../@/constants/quillConfig';
 import { call } from '../../@/shared';
 import Loader from '../../components/Loader';
 import { message, Alert } from '../../components/message';
 import { StateContext } from '../../LayoutContainer';
 import FormField from '../../components/FormField';
-import { defaultEmails, adminMenu } from '../../constants/general';
+import { defaultEmails, adminMenu } from '../../@/constants/general';
 
 function Emails({ history }) {
   const [loading, setLoading] = useState(true);
