+++ imports/ui/pages/admin/Members.jsx
@@ -24,7 +24,7 @@ import ListMenu from '../../components/ListMenu';
 import { message, Alert } from '../../components/message';
 import { StateContext } from '../../LayoutContainer';
 import { call } from '../../@/shared';
-import { adminMenu } from '../../constants/general';
+import { adminMenu } from '../../@/constants/general';
 import Hosts from '../../../api/@hosts/host';
 const compareUsersByDate = (a, b) => {
   const dateA = new Date(a.createdAt);
