+++ imports/ui/pages/profile/Profile.jsx
@@ -16,7 +16,7 @@ import ListMenu from '../../components/ListMenu';
 import Template from '../../components/Template';
 import ConfirmModal from '../../components/ConfirmModal';
 import { message } from '../../components/message';
-import { userMenu } from '../../constants/general';
+import { userMenu } from '../../@/constants/general';
 import { call, resizeImage, uploadImage } from '../../@/shared';
 import FileDropper from '../../components/FileDropper';
 import { StateContext } from '../../LayoutContainer';
