+++ imports/ui/pages/works/MyWorks.jsx
@@ -16,7 +16,7 @@ import Template from '../../components/Template';
 import ListMenu from '../../components/ListMenu';
 import Loader from '../../components/Loader';
 import { message, Alert } from '../../components/message';
-import { userMenu } from '../../constants/general';
+import { userMenu } from '../../@/constants/general';
 import WorkThumb from '../../components/WorkThumb';
 
 function Works({ history }) {
